--- modules/CMS/Abstracts/PluginServiceProvider.php
@@ -1,51 +1,0 @@
-<?php
-
-namespace Juzaweb\CMS\Abstracts;
-
-use Illuminate\Support\ServiceProvider;
-use Juzaweb\CMS\Contracts\PluginRepositoryInterface;
-use Juzaweb\CMS\Providers\BootstrapServiceProvider;
-use Juzaweb\CMS\Providers\ContractsServiceProvider;
-
-abstract class PluginServiceProvider extends ServiceProvider
-{
-    /**
-     * Register all plugins.
-     */
-    protected function registerModules()
-    {
-        $this->app->register(BootstrapServiceProvider::class);
-    }
-
-    /**
-     * Register package's namespaces.
-     */
-    protected function registerNamespaces()
-    {
-        $configPath = __DIR__ . '/../config/plugin.php';
-        $this->mergeConfigFrom($configPath, 'plugin');
-    }
-
-    /**
-     * Register the service provider.
-     */
-    abstract protected function registerServices();
-
-    /**
-     * Get the services provided by the provider.
-     *
-     * @return array
-     */
-    public function provides()
-    {
-        return [PluginRepositoryInterface::class, 'plugins'];
-    }
-
-    /**
-     * Register providers.
-     */
-    protected function registerProviders()
-    {
-        $this->app->register(ContractsServiceProvider::class);
-    }
-}
