+++ modules/CMS/Providers/PluginServiceProvider.php
@@ -2,13 +2,13 @@
 
 namespace Juzaweb\CMS\Providers;
 
-use Juzaweb\CMS\Abstracts\PluginServiceProvider as BaseServiceProvider;
 use Juzaweb\CMS\Contracts\ActivatorInterface;
 use Juzaweb\CMS\Contracts\PluginRepositoryInterface;
 use Juzaweb\CMS\Exceptions\InvalidActivatorClass;
 use Juzaweb\CMS\Support\LaravelFileRepository;
+use Juzaweb\CMS\Support\ServiceProvider;
 
-class PluginServiceProvider extends BaseServiceProvider
+class PluginServiceProvider extends ServiceProvider
 {
     /**
      * Booting the package.
@@ -29,8 +29,15 @@ class PluginServiceProvider extends BaseServiceProvider
     }
 
     /**
-     * {@inheritdoc}
+     * Get the services provided by the provider.
+     *
+     * @return array
      */
+    public function provides(): array
+    {
+        return [PluginRepositoryInterface::class, 'plugins'];
+    }
+
     protected function registerServices()
     {
         $this->app->singleton(
@@ -55,5 +62,30 @@ class PluginServiceProvider extends BaseServiceProvider
         );
 
         $this->app->alias(PluginRepositoryInterface::class, 'plugins');
+    }
+
+    /**
+     * Register all plugins.
+     */
+    protected function registerModules()
+    {
+        $this->app->register(BootstrapServiceProvider::class);
+    }
+
+    /**
+     * Register package's namespaces.
+     */
+    protected function registerNamespaces()
+    {
+        $configPath = __DIR__ . '/../config/plugin.php';
+        $this->mergeConfigFrom($configPath, 'plugin');
+    }
+
+    /**
+     * Register providers.
+     */
+    protected function registerProviders()
+    {
+        $this->app->register(ContractsServiceProvider::class);
     }
 }
